+++ mvp-widget-sample/src/main/java/ru/surfstudio/android/mvpwidget/sample/ui/screen/main/list/ListItemController.kt
@@ -19,7 +19,7 @@ class ListItemController(
         val constraintWidgetView = itemView.constraint_w
 
         override fun bind(int: Int) {
-            constraintWidgetView.widgetDataId = int.toString()
+            constraintWidgetView.widgetDataId = "id:${int}_"
             constraintWidgetView.render(int.toString())
         }
     }