+++ mvp-widget/src/main/java/ru/surfstudio/android/mvp/widget/view/CoreWidgetViewInterface.java
@@ -26,12 +26,9 @@ import ru.surfstudio.android.mvp.widget.scope.WidgetViewPersistentScope;
 /**
  * Интерфейс для всех кастомных вьюшек с презентером
  * <p>
- * !!!ВАЖНО!!!
- * <b/>При использовании в RecyclerView необходимо проставить атрибут enableManualInit == true в макете,
- * вызвать метод init(scopeId) во время bind()
- * <p>
+ * Для использования виджетов в RecyclerView, необходимо переопределить метод getWidgetId так,
+ * чтобы он получал значение из данных, получаемых в методе bind() у ViewHolder.
  */
-
 public interface CoreWidgetViewInterface extends
         PresenterHolderCoreView,
         HasConfigurator,
@@ -59,26 +56,13 @@ public interface CoreWidgetViewInterface extends
      * Для отложенной инициализации виджета с установкой идентификатора scope на основе данных
      * Используется, если у виджета проставлен атрибут - enableManualInit == true
      * <p>
-     * Необходимо использовать только в крайних случаях,
-     * когда невозможно проинициализировать виджет из верстки или при динамическом создании.
-     * (например, когда инициализация виджета происходит на основе данных, загружаемых из сети).
+     * Необходимо использовать только в крайних случаях, когда не устраивает автоматическая инициализация на onAttachedToWindow.
+     * Это может потребоваться в случае когда onAttachedToWindow вызывается до того, как становится известен widgetId.
      * <p>
      * Не рекомендуется использовать атрибут enableManualInit в связке с lazyInit
-     * для виджетов внутри RecyclerView. Для идентификации элементов в RecyclerView, необходимо переопределять метод getWidgetId.
+     * для виджетов внутри RecyclerView. Для идентификации элементов в RecyclerView, необходимо переопределять метод {@link CoreWidgetViewInterface#getWidgetId}.
      */
     void lazyInit();
 
     void onCreate();
-
-    /**
-     * Устрарел, для отложенной инициализации используется lazyInit
-     */
-    @Deprecated
-    void init();
-
-    /**
-     * Устрарел, для отложенной инициализации используется lazyInit
-     */
-    @Deprecated
-    void init(String scopeId);
 }